+++ test/test_task_short_read_assembler.py
@@ -38,13 +38,16 @@ def test_list_input_files():
         nose.assert_true(os.path.isfile(location))
 
 def test_copy_output_files():
-    app  = app_helper.mock_short_read_assembler_state(intermediates = True)
+    app = app_helper.mock_short_read_assembler_state(intermediates = True)
     run.copy_output_files(app)
-    file_helper.assert_is_file(fs.get_output_file_path('contig_fasta/7e9f760161', app))
+    file_helper.assert_is_file(fs.get_task_file_path(app, 'outputs/container_log/log.txt'))
+    file_helper.assert_is_file(fs.get_task_file_path(app, 'outputs/contig_fasta/7e9f760161'))
 
 
 @attr('slow')
 def test_complete_run_through():
     app = app_helper.mock_short_read_assembler_state(reads = True)
     run.execute_image(app)
-    file_helper.assert_is_file(fs.get_output_file_path('contig_fasta/7e9f760161', app))
+    file_helper.assert_is_file(fs.get_task_file_path(app, 'outputs/contig_fasta/7e9f760161'))
+    file_helper.assert_is_file(fs.get_task_file_path(app, 'outputs/container_runtime_metrics/metrics.json'))
+    file_helper.assert_is_file(fs.get_task_file_path(app, 'outputs/container_log/log.txt'))